+++ hoomd/md/IntegrationMethodTwoStep.cc
@@ -26,15 +26,12 @@ namespace md
 IntegrationMethodTwoStep::IntegrationMethodTwoStep(std::shared_ptr<SystemDefinition> sysdef,
                                                    std::shared_ptr<ParticleGroup> group)
     : m_sysdef(sysdef), m_group(group), m_pdata(m_sysdef->getParticleData()),
-      m_exec_conf(m_pdata->getExecConf()), m_aniso(false), m_deltaT(Scalar(0.0)),
-      m_valid_restart(false)
+      m_exec_conf(m_pdata->getExecConf()), m_aniso(false), m_deltaT(Scalar(0.0))
     {
     // sanity check
     assert(m_sysdef);
     assert(m_pdata);
     assert(m_group);
-
-    m_integrator_id = m_sysdef->getIntegratorData()->registerIntegrator();
     }
 
 /*! \param deltaT New time step to set
@@ -42,47 +39,6 @@ IntegrationMethodTwoStep::IntegrationMethodTwoStep(std::shared_ptr<SystemDefinit
 void IntegrationMethodTwoStep::setDeltaT(Scalar deltaT)
     {
     m_deltaT = deltaT;
-    }
-
-/*! \param v is the restart variables for the current integrator
-    \param type is the type of expected integrator type
-    \param nvariables is the expected number of variables
-
-    If the either the integrator type or number of variables does not match the
-    expected values, this function throws the appropriate warning and returns
-    "false."  Otherwise, the function returns true.
-*/
-bool IntegrationMethodTwoStep::restartInfoTestValid(const IntegratorVariables& v,
-                                                    std::string type,
-                                                    unsigned int nvariables)
-    {
-    bool good = true;
-    if (v.type == "")
-        good = false;
-    else if (v.type != type && v.type != "")
-        {
-        m_exec_conf->msg->warning()
-            << "Integrator #" << m_integrator_id << " type " << type << " does not match type ";
-        m_exec_conf->msg->warning() << v.type << " found in restart file. " << endl;
-        m_exec_conf->msg->warning()
-            << "Ensure that the integrator order is consistent for restarted simulations. " << endl;
-        m_exec_conf->msg->warning() << "Continuing while ignoring restart information..." << endl;
-        good = false;
-        }
-    else if (v.type == type)
-        {
-        if (v.variable.size() != nvariables)
-            {
-            m_exec_conf->msg->warning()
-                << "Integrator #" << m_integrator_id << " type " << type << endl;
-            m_exec_conf->msg->warning()
-                << "appears to contain bad or incomplete restart information. " << endl;
-            m_exec_conf->msg->warning()
-                << "Continuing while ignoring restart information..." << endl;
-            good = false;
-            }
-        }
-    return good;
     }
 
 /*! \param query_group Group over which to count (translational) degrees of freedom.